Introduce MongoCollection to store objects of any type.
MongoCollection is an abstraction of the underlying MongoClient. It's
the nature of MongoClient, that it can only deal with JsonObjects.
Dealing with objects of other types is quite hard, because it requires a
transformation between these objects and JsonObject. MongoCollection
helps to bundle all necessary logic at one single point.

Change-Id: Ifdf7d542eb302c5f248965272252332497896fb0